lightshotParserTor - not valid/requirements.txt
colorama==0.4.3
termcolor==1.1.0
httpx==0.13.3
stem==1.8.0
SocksiPy_branch==1.01
